[arc] Modularize arc
This splits up the ARC main.c into separate .c and .h files.
