+++ arc/src/arc_grove_lcd.c
@@ -1,0 +1,99 @@
+// Copyright (c) 2017, Intel Corporation.
+
+// Zephyr includes
+#include <zephyr.h>
+#include <device.h>
+#include <string.h>
+#include <display/grove_lcd.h>
+
+// ZJS includes
+#include "arc_common.h"
+#include "arc_grove_lcd.h"
+
+#define MAX_BUFFER_SIZE 256
+
+static struct device *glcd = NULL;
+static char str[MAX_BUFFER_SIZE];
+
+void arc_handle_glcd(struct zjs_ipm_message *msg)
+{
+    char *buffer;
+    u8_t r, g, b;
+    u32_t error_code = ERROR_IPM_NONE;
+
+    if (msg->type != TYPE_GLCD_INIT && !glcd) {
+        ERR_PRINT("Grove LCD device not found\n");
+        ipm_send_error(msg, ERROR_IPM_OPERATION_FAILED);
+        return;
+    }
+
+    switch (msg->type) {
+    case TYPE_GLCD_INIT:
+        if (!glcd) {
+            /* Initialize the Grove LCD */
+            glcd = device_get_binding(GROVE_LCD_NAME);
+
+            if (!glcd) {
+                error_code = ERROR_IPM_OPERATION_FAILED;
+            } else {
+                DBG_PRINT("Grove LCD initialized\n");
+            }
+        }
+        break;
+    case TYPE_GLCD_PRINT:
+        /* print to LCD screen */
+        buffer = msg->data.glcd.buffer;
+        if (!buffer) {
+            error_code = ERROR_IPM_INVALID_PARAMETER;
+            ERR_PRINT("buffer not found\n");
+        } else {
+            snprintf(str, MAX_BUFFER_SIZE, "%s", buffer);
+            glcd_print(glcd, str, strlen(str));
+            DBG_PRINT("Grove LCD print: %s\n", str);
+        }
+        break;
+    case TYPE_GLCD_CLEAR:
+        glcd_clear(glcd);
+        break;
+    case TYPE_GLCD_SET_CURSOR_POS:
+        glcd_cursor_pos_set(glcd, msg->data.glcd.col, msg->data.glcd.row);
+        break;
+    case TYPE_GLCD_SET_COLOR:
+        r = msg->data.glcd.color_r;
+        g = msg->data.glcd.color_g;
+        b = msg->data.glcd.color_b;
+        glcd_color_set(glcd, r, g, b);
+        break;
+    case TYPE_GLCD_SELECT_COLOR:
+        glcd_color_select(glcd, msg->data.glcd.value);
+        break;
+    case TYPE_GLCD_SET_FUNCTION:
+        glcd_function_set(glcd, msg->data.glcd.value);
+        break;
+    case TYPE_GLCD_GET_FUNCTION:
+        msg->data.glcd.value = glcd_function_get(glcd);
+        break;
+    case TYPE_GLCD_SET_DISPLAY_STATE:
+        glcd_display_state_set(glcd, msg->data.glcd.value);
+        break;
+    case TYPE_GLCD_GET_DISPLAY_STATE:
+        msg->data.glcd.value = glcd_display_state_get(glcd);
+        break;
+    case TYPE_GLCD_SET_INPUT_STATE:
+        glcd_input_state_set(glcd, msg->data.glcd.value);
+        break;
+    case TYPE_GLCD_GET_INPUT_STATE:
+        msg->data.glcd.value = glcd_input_state_get(glcd);
+        break;
+    default:
+        ERR_PRINT("unsupported grove lcd message type %u\n", msg->type);
+        error_code = ERROR_IPM_NOT_SUPPORTED;
+    }
+
+    if (error_code != ERROR_IPM_NONE) {
+        ipm_send_error(msg, error_code);
+        return;
+    }
+
+    ipm_send_msg(msg);
+}
