+++ arc/src/arc_grove_lcd.h
@@ -1,0 +1,8 @@
+// Copyright (c) 2017, Intel Corporation.
+
+#ifndef __arc_grove_lcd_h__
+#define __arc_grove_lcd_h__
+
+void arc_handle_glcd(struct zjs_ipm_message *msg);
+
+#endif  // __arc_grove_lcd_h__
