+++ arc/src/arc_i2c.h
@@ -1,0 +1,8 @@
+// Copyright (c) 2017, Intel Corporation.
+
+#ifndef __arc_i2c_h__
+#define __arc_i2c_h__
+
+void arc_handle_i2c(struct zjs_ipm_message *msg);
+
+#endif  // __arc_i2c_h__
