+++ arc/src/arc_pme.h
@@ -1,0 +1,8 @@
+// Copyright (c) 2017, Intel Corporation.
+
+#ifndef __arc_pme_h__
+#define __arc_pme_h__
+
+void arc_handle_pme(struct zjs_ipm_message *msg);
+
+#endif  // __arc_pme_h__
