+++ arc/src/zjs_aio_arc.json
@@ -6,5 +6,6 @@
     "zephyr_conf": {
         "arc": ["CONFIG_ADC=y"]
     },
+    "src": ["arc_aio.c"],
     "zjs_config": ["-DBUILD_MODULE_AIO"]
 }
