+++ arc/src/zjs_grove_lcd_arc.json
@@ -11,5 +11,6 @@
             "CONFIG_GROVE_LCD_RGB_INIT_PRIORITY=90"
         ]
     },
+    "src": ["arc_grove_lcd.c"],
     "zjs_config": ["-DBUILD_MODULE_GROVE_LCD"]
 }
