+++ arc/src/zjs_i2c_arc.json
@@ -8,6 +8,6 @@
             "CONFIG_I2C=y"
         ]
     },
-    "src": ["../../src/zjs_ipm.c", "../../src/zjs_i2c_handler.c"],
+    "src": ["arc_i2c.c", "../../src/zjs_ipm.c", "../../src/zjs_i2c_handler.c"],
     "zjs_config": ["-DBUILD_MODULE_I2C"]
 }
