+++ arc/src/zjs_sensor_accel_arc.json
@@ -1,7 +1,8 @@
 {
     "module": "sensor_accel_arc",
     "constructor": "Accelerometer",
-    "targets": ["arc"],
     "depends": ["sensor_arc"],
+    "targets": ["arc"],
+    "src": ["arc_sensor.c"],
     "zjs_config": ["-DBUILD_MODULE_SENSOR_ACCEL"]
 }
