+++ arc/src/zjs_sensor_light_arc.json
@@ -6,5 +6,6 @@
     "zephyr_conf": {
         "arc": ["CONFIG_ADC=y"]
     },
+    "src": ["arc_sensor.c"],
     "zjs_config": ["-DBUILD_MODULE_SENSOR_LIGHT"]
 }
