+++ arc/src/zjs_sensor_temp_arc.json
@@ -3,5 +3,6 @@
     "constructor": "TemperatureSensor",
     "depends": ["sensor_arc"],
     "targets": ["arc"],
+    "src": ["arc_sensor.c"],
     "zjs_config": ["-DBUILD_MODULE_SENSOR_TEMP"]
 }
